requirements.txt
#data 
pandas==2.1.4
numpy==1.26.4
pyarrow==15.0.2
duckdb==0.10.3
scikit-learn==1.4.2

#modeling 
xgboost==2.0.3
lightgbm==4.3.0
optuna==3.6.1
lifelines==0.27.8

# experiment tracking 
mlflow==2.12.2
dvc==3.51.2

# explainability 
shap==0.44.1

#api
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
httpx==0.27.0

# dashboard
streamlit==1.35.0
plotly==5.22.0

# llm 
openai==1.30.1

# testing
pytest==8.2.0
pytest-cov==5.0.0

#utilities
python-dotenv==1.0.1
pyyaml==6.0.1
jupyter==1.0.0
matplotlib==3.8.4
seaborn==0.13.2